extract inference optimization into layer
