+++ webapi/src/main/scala/org/knora/webapi/responders/v2/SearchResponderV2.scala
@@ -40,6 +40,7 @@ import org.knora.webapi.messages.util.search.gravsearch.mainquery.GravsearchMain
 import org.knora.webapi.messages.util.search.gravsearch.prequery.AbstractPrequeryGenerator
 import org.knora.webapi.messages.util.search.gravsearch.prequery.GravsearchToCountPrequeryTransformer
 import org.knora.webapi.messages.util.search.gravsearch.prequery.GravsearchToPrequeryTransformer
+import org.knora.webapi.messages.util.search.gravsearch.prequery.InferenceOptimizationService
 import org.knora.webapi.messages.util.search.gravsearch.types.GravsearchTypeInspectionUtil
 import org.knora.webapi.messages.util.search.gravsearch.types._
 import org.knora.webapi.messages.util.standoff.StandoffTagUtilV2
@@ -67,6 +68,7 @@ final case class SearchResponderV2Live(
   private val queryTraverser: QueryTraverser,
   private val sparqlTransformerLive: SparqlTransformerLive,
   private val gravsearchTypeInspectionRunner: GravsearchTypeInspectionRunner,
+  private val inferenceOptimizationService: InferenceOptimizationService,
   implicit private val stringFormatter: StringFormatter
 ) extends SearchResponderV2
     with MessageHandler
@@ -406,7 +408,8 @@ final case class SearchResponderV2Live(
           stringFormatter
         )
 
-      ontologiesForInferenceMaybe <- queryTraverser.getOntologiesRelevantForInference(inputQuery.whereClause)
+      ontologiesForInferenceMaybe <-
+        inferenceOptimizationService.getOntologiesRelevantForInference(inputQuery.whereClause)
 
       triplestoreSpecificCountQuery <- queryTraverser.transformSelectToSelect(
                                          inputQuery = nonTriplestoreSpecificPrequery,
@@ -468,7 +471,8 @@ final case class SearchResponderV2Live(
       // TODO: if the ORDER BY criterion is a property whose occurrence is not 1, then the logic does not work correctly
       // TODO: the ORDER BY criterion has to be included in a GROUP BY statement, returning more than one row if property occurs more than once
 
-      ontologiesForInferenceMaybe <- queryTraverser.getOntologiesRelevantForInference(inputQuery.whereClause)
+      ontologiesForInferenceMaybe <-
+        inferenceOptimizationService.getOntologiesRelevantForInference(inputQuery.whereClause)
 
       nonTriplestoreSpecificPrequery <-
         queryTraverser.transformConstructToSelect(
@@ -1073,6 +1077,7 @@ object SearchResponderV2Live {
       with QueryTraverser
       with SparqlTransformerLive
       with GravsearchTypeInspectionRunner
+      with InferenceOptimizationService
       with StringFormatter,
     Nothing,
     SearchResponderV2Live
@@ -1091,6 +1096,7 @@ object SearchResponderV2Live {
         stringFormatter              <- ZIO.service[StringFormatter]
         mr                           <- ZIO.service[MessageRelay]
         typeInspectionRunner         <- ZIO.service[GravsearchTypeInspectionRunner]
+        inferenceOptimizationService <- ZIO.service[InferenceOptimizationService]
         handler <- mr.subscribe(
                      new SearchResponderV2Live(
                        appConfig,
@@ -1103,6 +1109,7 @@ object SearchResponderV2Live {
                        queryTraverser,
                        sparqlTransformerLive,
                        typeInspectionRunner,
+                       inferenceOptimizationService,
                        stringFormatter
                      )
                    )
